+++ app/src/main/java/elite/intel/ai/brain/actions/handlers/query/BaseQueryAnalyzer.java
@@ -5,6 +5,7 @@ import elite.intel.ai.ApiFactory;
 import elite.intel.ai.brain.AIConstants;
 import elite.intel.ai.brain.AiAnalysisInterface;
 import elite.intel.ai.brain.actions.handlers.query.struct.AiData;
+import elite.intel.ws.WebSocketBroadcaster;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -18,6 +19,7 @@ public class BaseQueryAnalyzer {
         if(originalUserInput == null) {originalUserInput = "";}
         AiAnalysisInterface aiAnalysisInterface = ApiFactory.getInstance().getAnalysisEndpoint();
         JsonObject analysis = aiAnalysisInterface.analyzeData(originalUserInput, struct);
+        WebSocketBroadcaster.getInstance().broadcast(analysis);
 
         if (!analysis.has(AIConstants.PROPERTY_TEXT_TO_SPEECH_RESPONSE)) {
             analysis = GenericResponse.getInstance().genericResponse("LLM failed to process this request.");
@@ -28,6 +30,7 @@ public class BaseQueryAnalyzer {
     protected JsonObject process(String message) {
         JsonObject object = new JsonObject();
         object.addProperty(AIConstants.PROPERTY_TEXT_TO_SPEECH_RESPONSE, message);
+        WebSocketBroadcaster.getInstance().broadcast(object);
         return object;
     }
 }